CRUD TIPO DE DOCUMENTO

+++ src/main/java/com/cursojava/curso/service/TipoDocumentoServiceAPI.java
@@ -3,5 +3,5 @@ package com.cursojava.curso.service;
 import com.cursojava.curso.commons.GenericServiceAPI;
 import com.cursojava.curso.model.TipoDocumento;
 
-public interface TipoDocumentoServiceAPI extends GenericServiceAPI<TipoDocumento, Long> {
+public interface TipoDocumentoServiceAPI extends GenericServiceAPI<TipoDocumento, Integer> {
 }
